requirements.txt
pypng
pyqrcode
Pillow
os-sys

